Menambah fitur edit transaksi pending

+++ application/controllers/Penjualanku.php
@@ -73,6 +73,14 @@ class Penjualanku extends CI_Controller
 
     public function buat()
     {
+        if($this->session->userdata('activity_edit_sales') != NULL){
+            $this->cart->destroy();
+            $this->session->unset_userdata('activity_edit_sales');
+            redirect(base_url("penjualanku/buat"));
+        }
+        if($this->session->userdata('activity_create_sales') == NULL){
+            $this->session->set_userdata('activity_create_sales',true);
+        }
         $this->load->model('barang_m');
         $this->load->model('toko_m');
         $barang = $this->barang_m->read()->result_array();
@@ -178,6 +186,7 @@ class Penjualanku extends CI_Controller
                 );
                 $this->pembayaran_m->create($data_pembayaran);
                 $this->cart->destroy();
+                $this->session->unset_userdata('activity_create_sales');
                 $this->session->set_flashdata(array(
                     'status' => 1,
                     'message' => 'Berhasil Membuat Transaksi',
@@ -196,9 +205,49 @@ class Penjualanku extends CI_Controller
 
     public function edit($id_transaksi_sales)
     {
-		$this->cart->destroy();
-		$this->load->model('barang_m');
+        $this->load->model('barang_m');
         $this->load->model('toko_m');
+        $this->load->model('pembayaran_m');
+        $this->load->model('transaksi_sales_m');
+        $this->load->model('item_transaksi_m');
+        $item = $this->item_transaksi_m->read_full_where(array(
+            'item_transaksi.id_transaksi_sales' => $id_transaksi_sales
+        ))->result();
+        $transaksi = $this->transaksi_sales_m->read_where(array(
+            'id_transaksi_sales' => $id_transaksi_sales
+        ))->row_array();
+        if($transaksi['status'] != 'pending'){
+            redirect(base_url('penjualanku'));
+        }
+        // $this->cart->destroy();
+        // $this->session->unset_userdata('activity_edit_sales');
+        if($this->session->userdata('activity_edit_sales') == NULL){
+            if($this->session->userdata('activity_create_sales') == true){
+                $this->cart->destroy();
+                $this->session->unset_userdata('activity_create_sales');
+                redirect(base_url("penjualanku/edit/$id_transaksi_sales"));
+            }
+            $this->session->set_userdata('activity_edit_sales',$id_transaksi_sales);
+            foreach($item as $data_barang){
+                $data_barang->potongan_harga = $data_barang->subdiskon;
+                $data_barang->sebelum_total = $data_barang->subtotal;
+                $data_input = array(
+                    'id' => $data_barang->id_barang,
+                    'qty' => (int)$data_barang->kuantitas + $data_barang->kuantitas_perbox,
+                    'price' => (int)$data_barang->harga + $data_barang->harga_perbox,
+                    'name' => $data_barang->nama_barang,
+                    'options' => $data_barang
+                );
+                $this->cart->insert($data_input);
+            }
+        } else {
+            if($this->session->userdata('activity_edit_sales') != $id_transaksi_sales){
+                $this->cart->destroy();
+                $this->session->unset_userdata('activity_edit_sales');
+                redirect(base_url("penjualanku/edit/$id_transaksi_sales"));
+            }
+        }
+		
         $barang = $this->barang_m->read()->result_array();
         $toko = $this->toko_m->read()->result_array();
         $potongan_harga = 0;
@@ -213,6 +262,10 @@ class Penjualanku extends CI_Controller
                 $potongan_harga += $this->cart->product_options($item['rowid'])->potongan_harga;
             }
         }
+        $pembayaran = $this->pembayaran_m->read_where(array(
+            'id_transaksi_sales' => $id_transaksi_sales
+        ))->row_array();
+      
         $data = array(
             'konten' => 'sales/form_penjualan',
             'parsing' => array(
@@ -220,15 +273,127 @@ class Penjualanku extends CI_Controller
                 'toko' => $toko,
                 'potongan_harga' => $potongan_harga,
                 'sebelum_total' => $sebelum_total,
+                'toko_dipilih' => $transaksi['id_toko'],
+                'pembayaran' => $pembayaran['jumlah_pembayaran'],
+                'id_transaksi_sales' => $id_transaksi_sales,
+                'id_toko_dipilih' => $transaksi['id_toko']
             ),
             'js' => 'sales/js_penjualan',
         );
         $this->load->view('_partials/template', $data);
 	}
 	
-	public function aksi_edit()
+	public function aksi_edit($id_transaksi_sales)
 	{
+        $this->load->model('transaksi_sales_m');
+        $this->load->model('item_transaksi_m');
+        $this->load->model('barang_m');
+        $this->load->model('pembayaran_m');
+        if ($this->cart->contents() == null) {
+            $this->session->set_flashdata(array(
+                'status' => 0,
+                'message' => 'Tambahkan minimal satu barang',
+            ));
+            redirect(base_url("penjualanku/edit/$id_transaksi_sales"));
+        } else {
+            $data_transaksi = $this->input->post();
+            $id_sales = $this->session->userdata('id');
 
+            $stok_habis = false;
+            $stok_habis_perbox = false;
+            $pengajuan_pcs = false;
+            $pengajuan_box = false;
+            $jumlah_diskon = 0;
+            $sebelum_total = 0;
+            foreach ($this->cart->contents() as $item) {
+                $barang = $this->barang_m->read_where(array('id_barang' => $item['id']))->row();
+                if ($barang->stok < $this->cart->product_options($item['rowid'])->kuantitas) {
+                    $stok_habis = true;
+                } else if ($barang->stok_perbox < $this->cart->product_options($item['rowid'])->kuantitas_perbox) {
+                    $stok_habis_perbox = true;
+                }
+                $jumlah_diskon += $this->cart->product_options($item['rowid'])->potongan_harga;
+                $sebelum_total += $this->cart->product_options($item['rowid'])->sebelum_total;
+                if ($this->cart->product_options($item['rowid'])->pengajuan_pcs > 0) {
+                    $pengajuan_pcs = true;
+                }
+                if ($this->cart->product_options($item['rowid'])->pengajuan_box > 0) {
+                    $pengajuan_box = true;
+                }
+            }
+
+            if (!$stok_habis && !$stok_habis_perbox) {
+                $data_input = array(
+                    'id_sales' => $id_sales,
+                    'id_toko' => $data_transaksi['id_toko'],
+                    'total' => $sebelum_total - $jumlah_diskon,
+                    'diskon' => $jumlah_diskon,
+                );
+                if ($this->priceToFloat($data_transaksi['pembayaran']) == $data_input['total']) {
+                    $data_input['is_lunas'] = 1;
+                }
+                if ($pengajuan_pcs == true || $pengajuan_box == true) {
+                    $data_input['status'] = 'pending';
+                }
+                $this->transaksi_sales_m->update($data_input,array('id_transaksi_sales' => $id_transaksi_sales));
+                $item_hapus = $this->item_transaksi_m->read_where(array(
+                    'id_transaksi_sales' => $id_transaksi_sales
+                ))->result_array();
+                foreach($item_hapus as $item){
+                    $barang = $this->barang_m->read_where(array('id_barang' => $item['id_barang']))->row();
+                    $this->barang_m->update(array(
+                        'stok' => $barang->stok + $item['stok'],
+                        'stok_perbox' => $barang->stok_perbox + $item['stok_perbox']
+                    ), array(
+                        'id_barang' => $barang->id_barang
+                    ));
+                }
+                $this->item_transaksi_m->delete(array(
+                    'id_transaksi_sales' => $id_transaksi_sales 
+                ));
+                foreach ($this->cart->contents() as $item) {
+                    $barang = $this->barang_m->read_where(array('id_barang' => $item['id']))->row();
+                    $item_input = array(
+                        'id_transaksi_sales' => $id_transaksi_sales,
+                        'id_barang' => $item['id'],
+                        'kuantitas' => $this->cart->product_options($item['rowid'])->kuantitas,
+                        'kuantitas_perbox' => $this->cart->product_options($item['rowid'])->kuantitas_perbox,
+                        'harga_fix_pcs' => $this->cart->product_options($item['rowid'])->harga,
+                        'harga_fix_box' => $this->cart->product_options($item['rowid'])->harga_perbox,
+                        'subtotal' => $this->cart->product_options($item['rowid'])->sebelum_total,
+                        'subdiskon' => $this->cart->product_options($item['rowid'])->potongan_harga,
+                    );
+                    $this->item_transaksi_m->create($item_input);
+                    $this->barang_m->update(array(
+                        'stok' => $barang->stok - $this->cart->product_options($item['rowid'])->kuantitas,
+                        'stok_perbox' => $barang->stok_perbox - $this->cart->product_options($item['rowid'])->kuantitas_perbox,
+                    ), array(
+                        'id_barang' => $item['id'],
+                    ));
+                }
+                $data_pembayaran = array(
+                    'jumlah_pembayaran' => $this->priceToFloat($data_transaksi['pembayaran']),
+                );
+                $this->pembayaran_m->update($data_pembayaran,array(
+                    'id_transaksi_sales' => $id_transaksi_sales
+                ));
+                $this->cart->destroy();
+                $this->session->unset_userdata('activity_edit_sales');
+                $this->session->set_flashdata(array(
+                    'status' => 1,
+                    'message' => 'Berhasil Mengedit Transaksi',
+                ));
+                redirect(base_url("penjualanku/toko/$data_transaksi[id_toko]"));
+            } else {
+                $this->cart->destroy();
+                $this->session->unset_userdata('activity_edit_sales');
+                $this->session->set_flashdata(array(
+                    'status' => 0,
+                    'message' => 'Stok barang tidak mencukupi, silahkan ulangi',
+                ));
+                redirect(base_url("penjualanku/edit/$id_transaksi_sales"));
+            }
+        }
 	}
 
     public function detail($id_transaksi)
@@ -319,5 +484,6 @@ class Penjualanku extends CI_Controller
     public function test()
     {
         var_dump($this->cart->contents());
+        echo $this->session->userdata('activity_edit_sales');
     }
 }
